Add option to specify a base package to scan for domains
In a unit test, a list of domain classes to mock can be specified. That
initial list is then resolved into a full list including any required
objects, subclasses, and parent classes. If subclasses for an abstract
domain need to be found, the entire classpath is searched for components
assignable from any given abstract parent class. This is very expensive,
so to improve performance, a base package can be set to scan from, thus
reducing the time to scan for subclasses.